Song/Dance Buff Refresh Behavior (Attempt 1) - Song/Dance buff durations can no longer be refreshed once they become finite   * When the song/dance buff expires while in the area, it will get restored at that time   * Added a config to restore previous behavior - Fixed an issue that the "onleft" event due to dissonance was only invoked on the moving song/dance   * Now it will also be invoked for the songs/dances that it moved onto - Fixed the song/dance buff status icon to show a fake duration   * Now the duration will show as unlimited when still in the song/dance area   * The duration on the icon will be updated to the real duration once the song/dance was left - Fixes #1031

+++ src/map/skill.cpp
@@ -15780,6 +15780,8 @@ static int32 skill_dance_overlap_sub(struct block_list* bl, va_list ap)
 				// Add ensemble to signal this unit is overlapping.
 				unit->val2 |= (1 << UF_ENSEMBLE);
 				skill_getareachar_skillunit_visibilty(unit, AREA);
+				// Trigger onleft event on all affected units on the cell
+				map_foreachincell(skill_unit_effect, unit->bl.m, unit->bl.x, unit->bl.y, unit->group->bl_flag, &unit->bl, gettick(), 4|8);
 			}
 		}
 		else {
@@ -16691,7 +16693,7 @@ static int32 skill_unit_onplace(struct skill_unit *unit, struct block_list *bl, 
 			if (!sc) return 0;
 			if (!sce)
 				sc_start4(ss, bl,type,100,sg->skill_lv,sg->val1,sg->val2,0,sg->limit);
-			else if (sce->val4 == 1) { //Readjust timers since the effect will not last long.
+			else if (sce->val4 == 1 && battle_config.refresh_song == 1) { //Readjust timers since the effect will not last long.
 				sce->val4 = 0; //remove the mark that we stepped out
 				delete_timer(sce->timer, status_change_timer);
 				sce->timer = add_timer(tick+sg->limit, status_change_timer, bl->id, type); //put duration back to 3min
@@ -17871,13 +17873,17 @@ int32 skill_unit_onleft(uint16 skill_id, struct block_list *bl, t_tick tick)
 		case DC_DONTFORGETME:
 		case DC_FORTUNEKISS:
 		case DC_SERVICEFORYOU:
-			if (sce)
+			if (sce && (sce->val4 == 0 || battle_config.refresh_song == 1))
 			{
 				delete_timer(sce->timer, status_change_timer);
 				//NOTE: It'd be nice if we could get the skill_lv for a more accurate extra time, but alas...
 				//not possible on our current implementation.
+				int32 duration = skill_get_time2(skill_id, 1);
 				sce->val4 = 1; //Store the fact that this is a "reduced" duration effect.
-				sce->timer = add_timer(tick+skill_get_time2(skill_id,1), status_change_timer, bl->id, type);
+				sce->timer = add_timer(tick + duration, status_change_timer, bl->id, type);
+				// Update icon duration
+				std::shared_ptr<s_status_change_db> scdb = status_db.find(type);
+				clif_status_change(bl, static_cast<int32>(scdb->icon), 1, duration, 1, 0, 0);
 			}
 			break;
 		case PF_FOGWALL:
@@ -17912,7 +17918,7 @@ int32 skill_unit_onleft(uint16 skill_id, struct block_list *bl, t_tick tick)
  * flag values:
  * flag&1: Invoke onplace function (otherwise invoke onout)
  * flag&4: Invoke a onleft call (the unit might be scheduled for deletion)
- * flag&8: Recursive
+ * flag&8: Skip dissonance check
  *------------------------------------------*/
 static int32 skill_unit_effect(struct block_list* bl, va_list ap)
 {
@@ -17931,11 +17937,11 @@ static int32 skill_unit_effect(struct block_list* bl, va_list ap)
 	if (group == nullptr)
 		return 0;
 
-	if( !(flag&8) ) {
+	if( !(flag&8) )
 		dissonance = skill_dance_switch(unit, false);
-		//Target-type check.
-		isTarget = group->bl_flag & bl->type && battle_check_target( &unit->bl, bl, group->target_flag ) > 0;
-	}
+
+	//Target-type check.
+	isTarget = group->bl_flag & bl->type && battle_check_target(&unit->bl, bl, group->target_flag) > 0;
 
 	//Necessary in case the group is deleted after calling on_place/on_out [Skotlex]
 	skill_id = group->skill_id;
@@ -17952,11 +17958,8 @@ static int32 skill_unit_effect(struct block_list* bl, va_list ap)
 	} else if( !isTarget && flag&4 && ( group->state.song_dance&0x1 || ( group->src_id == bl->id && group->state.song_dance&0x2 ) ) )
 		skill_unit_onleft(skill_id, bl, tick);//Ensemble check to terminate it.
 
-	if( dissonance ) {
+	if( dissonance )
 		skill_dance_switch(unit, true);
-		//we placed a dissonance, let's update
-		map_foreachincell(skill_unit_effect,unit->bl.m,unit->bl.x,unit->bl.y,group->bl_flag,&unit->bl,gettick(),4|8);
-	}
 
 	return 0;
 }
